get-orig-source: conditional cleanup.

+++ debian/get-orig-source.sh
@@ -72,7 +72,9 @@ drop_files_excluded() {
         perl -0nE 'say $1 if m{^Files\-Excluded:\s*(.*?)(?:\n\n|^Files|^Comment)}sm;' debian/copyright \
         | ( cd "${work_dir}" && xargs --no-run-if-empty rm -rf )
     done
-    find "${work_dir}"/vendor -mindepth 1 -type d -empty -delete -printf 'removed %p\n'
+    if [ -d "${work_dir}"/vendor ]; then
+        find "${work_dir}"/vendor -mindepth 1 -type d -empty -delete -printf 'removed %p\n'
+    fi
 }
 
 ## extract main tarball:
